Added Messages and Commands.
Commands represented by INDEX | REMOVE | QUERY are now represented in
the type system. A type Message was introduced to represent the
combination of a Command + package the command is related to + any
dependencies that package has.

In addition, Message introduced a Decoding method that returns a Try
parameterized on the Message itself.